requirements.txt
pandas
numpy
python-dotenv
mysql-connector-python
seaborn
matplotlib
PyMySQL
scikit-learn
catboost
xgboost
#-e .
